Moving Chrome Custom tabs into separate module
