+++ app/src/main/java/com/jraska/github/client/AppModule.kt
@@ -3,23 +3,15 @@ package com.jraska.github.client
 import android.app.Application
 import android.app.NotificationManager
 import android.content.Context
-import android.view.LayoutInflater
 import com.facebook.drawee.backends.pipeline.Fresco
 import com.jakewharton.threetenabp.AndroidThreeTen
 import com.jraska.github.client.core.android.OnAppCreate
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.IntoSet
-import javax.inject.Singleton
 
 @Module
 object AppModule {
-
-  @Provides
-  @Singleton
-  internal fun provideLayoutInflater(context: Context): LayoutInflater {
-    return LayoutInflater.from(context)
-  }
 
   @Provides
   internal fun notificationManager(context: Context): NotificationManager {
