+++ feature/chrome-custom-tabs/build.gradle
@@ -1,0 +1,26 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+
+android {
+  compileSdkVersion 29
+  defaultConfig {
+    minSdkVersion 21
+  }
+
+  compileOptions {
+    sourceCompatibility JavaVersion.VERSION_1_8
+    targetCompatibility JavaVersion.VERSION_1_8
+  }
+}
+
+dependencies {
+  implementation project(':core')
+  implementation project(':core-android')
+
+  implementation 'com.google.dagger:dagger:2.27'
+  kapt 'com.google.dagger:dagger-compiler:2.27'
+  implementation 'androidx.browser:browser:1.2.0'
+
+  testImplementation 'junit:junit:4.13'
+}
